Thank you for your contribution to Spitball repo.
Before submitting this PR, please make sure:

- [ ] Your code builds clean without any errors
- [ ] You ran lint command on npm

Related work items: #3352, #3359, #3401, #3409, #3444, #3463, #3465, #3491, #3493, #3496, #3497, #3498

+++ Cloudents.Admin2/Models/PaymentRequest.cs
@@ -6,9 +6,9 @@ namespace Cloudents.Admin2.Models
 {
     public class PaymentRequest : IValidatableObject
     {
-        public decimal StudentPay { get; set; }
+        public double StudentPay { get; set; }
         //[Range(5, 50000)]
-        public decimal SpitballPay { get; set; }
+        public double SpitballPay { get; set; }
 
         public long UserId { get; set; }
         public long TutorId { get; set; }
